Keep the text wrapping/truncation and max lines stuff in Controls for backward compatibility Leave room for a better design in Core
