+++ src/Controls/src/Core/HandlerImpl/Button/Button.Android.cs
@@ -11,5 +11,10 @@ namespace Microsoft.Maui.Controls
 		{
 			handler.PlatformView?.UpdateText(button);
 		}
+
+		public static void MapLineBreakMode(IButtonHandler handler, Button button)
+		{
+			handler.PlatformView?.UpdateLineBreakMode(button);
+		}
 	}
 }
