+++ src/Controls/src/Core/HandlerImpl/Button/Button.Windows.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using Microsoft.Maui.Controls.Internals;
+using Microsoft.Maui.Controls.Platform;
 using Microsoft.Maui.Handlers;
 
 namespace Microsoft.Maui.Controls
@@ -19,6 +20,11 @@ namespace Microsoft.Maui.Controls
 			var text = TextTransformUtilites.GetTransformedText(button.Text, button.TextTransform);
 			handler.PlatformView?.UpdateText(text);
 			button.Handler?.UpdateValue(nameof(Button.ContentLayout));
+		}
+
+		public static void MapLineBreakMode(IButtonHandler handler, Button button)
+		{
+			handler.PlatformView?.UpdateLineBreakMode(button);
 		}
 	}
 }
