+++ src/Controls/src/Core/HandlerImpl/Button/Button.cs
@@ -24,6 +24,7 @@ namespace Microsoft.Maui.Controls
 #endif
 			[nameof(TextTransform)] = MapText,
 			[nameof(Text)] = MapText,
+			[nameof(Button.LineBreakMode)] = MapLineBreakMode,
 		};
 
 		internal new static void RemapForControls()
