+++ src/Controls/src/Core/HandlerImpl/Label/Label.Android.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Maui.Handlers;
+using Microsoft.Maui.Controls.Platform;
 
 namespace Microsoft.Maui.Controls
 {
@@ -12,6 +13,16 @@ namespace Microsoft.Maui.Controls
 		public static void MapText(LabelHandler handler, Label label)
 		{
 			Platform.TextViewExtensions.UpdateText(handler.PlatformView, label);
+		}
+
+		public static void MapLineBreakMode(LabelHandler handler, Label label)
+		{
+			handler.PlatformView?.UpdateLineBreakMode(label);
+		}
+
+		public static void MapMaxLines(ILabelHandler handler, Label label)
+		{
+			handler.PlatformView?.UpdateMaxLines(label);
 		}
 	}
 }
