+++ src/Controls/src/Core/HandlerImpl/Label/Label.Standard.cs
@@ -9,5 +9,8 @@ namespace Microsoft.Maui.Controls
 		public static void MapTextType(LabelHandler handler, Label label) { }
 		/// <include file="../../../../docs/Microsoft.Maui.Controls/Label.xml" path="//Member[@MemberName='MapText']/Docs" />
 		public static void MapText(LabelHandler handler, Label label) { }
+
+		public static void MapLineBreakMode(LabelHandler handler, Label label) { }
+		public static void MapMaxLines(LabelHandler handler, Label label) { }
 	}
 }
