+++ src/Controls/src/Core/HandlerImpl/Label/Label.Windows.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using Microsoft.Maui.Handlers;
+using Microsoft.Maui.Controls.Platform;
 
 namespace Microsoft.Maui.Controls
 {
@@ -12,5 +13,11 @@ namespace Microsoft.Maui.Controls
 
 		public static void MapText(LabelHandler handler, Label label) =>
 			Platform.TextBlockExtensions.UpdateText(handler.PlatformView, label);
+
+		public static void MapLineBreakMode(ILabelHandler handler, Label label) =>
+			handler.PlatformView?.UpdateLineBreakMode(label.LineBreakMode);
+
+		public static void MapMaxLines(ILabelHandler handler, Label label) =>
+			handler.PlatformView?.UpdateMaxLines(label);
 	}
 }
