+++ src/Controls/src/Core/Platform/iOS/Extensions/TextExtensions.cs
@@ -18,5 +18,50 @@ namespace Microsoft.Maui.Controls.Platform
 		{
 			textField.Text = TextTransformUtilites.GetTransformedText(inputView.Text, inputView.TextTransform);
 		}
+
+		public static void UpdateLineBreakMode(this UILabel platformLabel, Label label)
+		{
+			platformLabel.SetLineBreakMode(label);
+		}
+
+		public static void UpdateMaxLines(this UILabel platformLabel, Label label)
+		{
+			platformLabel.SetLineBreakMode(label);
+		}
+
+		internal static void SetLineBreakMode(this UILabel platformLabel, Label label)
+		{
+			int maxLines = label.MaxLines;
+			if (maxLines < 0)
+				maxLines = 0;
+
+			switch (label.LineBreakMode)
+			{
+				case LineBreakMode.NoWrap:
+					platformLabel.LineBreakMode = UILineBreakMode.Clip;
+					maxLines = 1;
+					break;
+				case LineBreakMode.WordWrap:
+					platformLabel.LineBreakMode = UILineBreakMode.WordWrap;
+					break;
+				case LineBreakMode.CharacterWrap:
+					platformLabel.LineBreakMode = UILineBreakMode.CharacterWrap;
+					break;
+				case LineBreakMode.HeadTruncation:
+					platformLabel.LineBreakMode = UILineBreakMode.HeadTruncation;
+					maxLines = 1;
+					break;
+				case LineBreakMode.MiddleTruncation:
+					platformLabel.LineBreakMode = UILineBreakMode.MiddleTruncation;
+					maxLines = 1;
+					break;
+				case LineBreakMode.TailTruncation:
+					platformLabel.LineBreakMode = UILineBreakMode.TailTruncation;
+					maxLines = 1;
+					break;
+			}
+
+			platformLabel.Lines = maxLines;
+		}
 	}
 }
