--- src/Core/src/Core/ILineBreakMode.cs
@@ -1,10 +1,0 @@
-﻿namespace Microsoft.Maui
-{
-	public interface ILineBreakMode
-	{
-		/// <summary>
-		/// Gets the option for line breaking.
-		/// </summary>
-		LineBreakMode LineBreakMode { get; }
-	}
-}
