+++ src/Core/src/Core/ITextButton.cs
@@ -3,7 +3,7 @@
 	/// <summary>
 	/// Functionality related with the Button View text.
 	/// </summary>
-	public interface ITextButton : IView, IButton, IText, ILineBreakMode
+	public interface ITextButton : IView, IButton, IText
 	{
 	}
 }
