+++ src/Core/src/Handlers/Button/ButtonHandler.Android.cs
@@ -111,12 +111,6 @@ namespace Microsoft.Maui.Handlers
 			handler.PlatformView?.UpdatePadding(button, DefaultPadding);
 		}
 
-		public static void MapLineBreakMode(IButtonHandler handler, IButton button)
-		{
-			if (button is ILineBreakMode lineBreakMode)
-				handler.PlatformView?.UpdateLineBreakMode(lineBreakMode);
-		}
-
 		public static void MapImageSource(IButtonHandler handler, IImage image) =>
 			MapImageSourceAsync(handler, image).FireAndForget(handler);
