+++ src/Core/src/Handlers/Button/ButtonHandler.Standard.cs
@@ -14,7 +14,6 @@ namespace Microsoft.Maui.Handlers
 		public static void MapCharacterSpacing(IButtonHandler handler, ITextStyle button) { }
 		public static void MapFont(IButtonHandler handler, ITextStyle button) { }
 		public static void MapPadding(IButtonHandler handler, IButton button) { }
-		public static void MapLineBreakMode(IButtonHandler handler, IButton button) { }
 		public static void MapImageSource(IButtonHandler handler, IImage image) { }
 
 		void OnSetImageSource(object? obj) { }
