+++ src/Core/src/Handlers/Button/ButtonHandler.Windows.cs
@@ -116,12 +116,6 @@ namespace Microsoft.Maui.Handlers
 			handler.PlatformView?.UpdatePadding(button);
 		}
 
-		public static void MapLineBreakMode(IButtonHandler handler, IButton button)
-		{
-			if (button is ILineBreakMode lineBreakMode)
-				handler.PlatformView?.UpdateLineBreakMode(lineBreakMode);
-		}
-
 		public static void MapImageSource(IButtonHandler handler, IImage image) =>
 			handler
 				.ImageSourceLoader
