+++ src/Core/src/Handlers/Label/LabelHandler.Android.cs
@@ -62,16 +62,6 @@ namespace Microsoft.Maui.Handlers
 			handler.PlatformView?.UpdateVerticalTextAlignment(label);
 		}
 
-		public static void MapLineBreakMode(ILabelHandler handler, ILabel label)
-		{
-			handler.PlatformView?.UpdateLineBreakMode(label);
-		}
-
-		public static void MapMaxLines(ILabelHandler handler, ILabel label)
-		{
-			handler.PlatformView?.UpdateMaxLines(label);
-		}
-
 		public static void MapPadding(ILabelHandler handler, ILabel label)
 		{
 			handler.PlatformView?.UpdatePadding(label);
