+++ src/Core/src/Handlers/Label/LabelHandler.Standard.cs
@@ -12,7 +12,6 @@ namespace Microsoft.Maui.Handlers
 		public static void MapFont(ILabelHandler handler, ILabel label) { }
 		public static void MapHorizontalTextAlignment(ILabelHandler handler, ILabel label) { }
 		public static void MapVerticalTextAlignment(ILabelHandler handler, ILabel label) { }
-		public static void MapLineBreakMode(ILabelHandler handler, ILabel label) { }
 		public static void MapTextDecorations(ILabelHandler handler, ILabel label) { }
 		public static void MapMaxLines(ILabelHandler handler, ILabel label) { }
 		public static void MapPadding(ILabelHandler handler, ILabel label) { }
