+++ src/Core/src/Handlers/Label/LabelHandler.cs
@@ -23,9 +23,7 @@ namespace Microsoft.Maui.Handlers
 			[nameof(ITextStyle.Font)] = MapFont,
 			[nameof(ITextAlignment.HorizontalTextAlignment)] = MapHorizontalTextAlignment,
 			[nameof(ITextAlignment.VerticalTextAlignment)] = MapVerticalTextAlignment,
-			[nameof(ILabel.LineBreakMode)] = MapLineBreakMode,
 			[nameof(ILabel.LineHeight)] = MapLineHeight,
-			[nameof(ILabel.MaxLines)] = MapMaxLines,
 			[nameof(ILabel.Padding)] = MapPadding,
 			[nameof(ILabel.Text)] = MapText,
 			[nameof(ITextStyle.TextColor)] = MapTextColor,
