+++ src/Core/src/Platform/Android/ButtonExtensions.cs
@@ -44,10 +44,5 @@ namespace Microsoft.Maui.Platform
 				(int)padding.Right,
 				(int)padding.Bottom);
 		}
-	
-		public static void UpdateLineBreakMode(this Button nativeControl, ILineBreakMode button)
-		{
-			nativeControl.SetLineBreakMode(button);
-		}
 	}
 }
