+++ src/Core/src/Platform/Windows/ButtonExtensions.cs
@@ -130,14 +130,6 @@ namespace Microsoft.Maui.Platform
 		public static void UpdatePadding(this Button platformButton, IPadding padding) =>
 			platformButton.UpdatePadding(padding, platformButton.GetResource<UI.Xaml.Thickness>("ButtonPadding"));
 
-		public static void UpdateLineBreakMode(this Button platformButton, ILineBreakMode button)
-		{
-			if (platformButton.GetContent<TextBlock>() is TextBlock textBlock)
-			{
-				textBlock?.UpdateLineBreakMode(button);
-			}
-		}
-    
 		public static void UpdateCharacterSpacing(this ButtonBase platformButton, ITextStyle button)
 		{
 			var characterSpacing = button.CharacterSpacing.ToEm();
