+++ src/Core/src/Platform/Windows/TextBlockExtensions.cs
@@ -37,18 +37,6 @@ namespace Microsoft.Maui.Platform
 			platformControl.CharacterSpacing = label.CharacterSpacing.ToEm();
 		}
 
-		public static void UpdateMaxLines(this TextBlock platformControl, ILabel label)
-		{
-			if (label.MaxLines >= 0)
-			{
-				platformControl.MaxLines = label.MaxLines;
-			}
-			else
-			{
-				platformControl.MaxLines = 0;
-			}
-		}
-
 		public static void UpdateTextDecorations(this TextBlock platformControl, ILabel label)
 		{
 			var elementTextDecorations = label.TextDecorations;
@@ -132,53 +120,6 @@ namespace Microsoft.Maui.Platform
 				platformControl.Text = global::Windows.Data.Html.HtmlUtilities.ConvertToText(label.Text);
 			}
 		}
-
-		public static void UpdateLineBreakMode(this TextBlock platformControl, ILineBreakMode breakMode)
-		{
-			var lineBreakMode = breakMode.LineBreakMode;
-			var label = breakMode as ILabel;
-
-			switch (lineBreakMode)
-			{
-				case LineBreakMode.NoWrap:
-					platformControl.TextTrimming = TextTrimming.Clip;
-					platformControl.TextWrapping = TextWrapping.NoWrap;
-					break;
-				case LineBreakMode.WordWrap:
-					platformControl.TextTrimming = TextTrimming.None;
-					platformControl.TextWrapping = TextWrapping.Wrap;
-					break;
-				case LineBreakMode.CharacterWrap:
-					platformControl.TextTrimming = TextTrimming.WordEllipsis;
-					platformControl.TextWrapping = TextWrapping.Wrap;
-					break;
-				case LineBreakMode.HeadTruncation:
-					// TODO: This truncates at the end.
-					platformControl.TextTrimming = TextTrimming.WordEllipsis;
-
-					if (label != null)
-						platformControl.DetermineTruncatedTextWrapping(label);
-					break;
-				case LineBreakMode.TailTruncation:
-					platformControl.TextTrimming = TextTrimming.CharacterEllipsis;
-
-					if (label != null)
-						platformControl.DetermineTruncatedTextWrapping(label);
-					break;
-				case LineBreakMode.MiddleTruncation:
-					// TODO: This truncates at the end.
-					platformControl.TextTrimming = TextTrimming.WordEllipsis;
-
-					if (label != null)
-						platformControl.DetermineTruncatedTextWrapping(label);
-					break;
-				default:
-					throw new ArgumentOutOfRangeException();
-			}
-		}
-
-		internal static void DetermineTruncatedTextWrapping(this TextBlock textBlock, ILabel label) =>
-			textBlock.TextWrapping = label.MaxLines > 1 ? TextWrapping.Wrap : TextWrapping.NoWrap;
 
 		internal static void UpdateTextPlainText(this TextBlock platformControl, IText label)
 		{
