+++ src/Core/src/Platform/iOS/ButtonExtensions.cs
@@ -76,18 +76,6 @@ namespace Microsoft.Maui.Platform
 				(float)padding.Right);
 		}
 
-		public static void UpdateLineBreakMode(this UIButton nativeButton, ILineBreakMode button)
-		{
-			nativeButton.TitleLabel.LineBreakMode = button.LineBreakMode switch
-			{
-				LineBreakMode.NoWrap => UILineBreakMode.Clip,
-				LineBreakMode.WordWrap => UILineBreakMode.WordWrap,
-				LineBreakMode.CharacterWrap => UILineBreakMode.CharacterWrap,
-				LineBreakMode.HeadTruncation => UILineBreakMode.HeadTruncation,
-				LineBreakMode.TailTruncation => UILineBreakMode.TailTruncation,
-				LineBreakMode.MiddleTruncation => UILineBreakMode.MiddleTruncation,
-				_ => throw new ArgumentOutOfRangeException()
-			};
-		}
+		
 	}
 }
