+++ src/Core/tests/DeviceTests/Handlers/Label/LabelHandlerTests.cs
@@ -248,129 +248,6 @@ namespace Microsoft.Maui.DeviceTests
 				() => label.TextDecorations = newDecorations);
 		}
 
-		[Fact(DisplayName = "LineBreakMode Initializes Correctly")]
-		public async Task LineBreakModeInitializesCorrectly()
-		{
-			var xplatLineBreakMode = LineBreakMode.TailTruncation;
-
-			var label = new LabelStub()
-			{
-				LineBreakMode = xplatLineBreakMode
-			};
-
-			var expectedValue = xplatLineBreakMode.ToPlatform();
-
-			var values = await GetValueAsync(label, (handler) =>
-			{
-				return new
-				{
-					ViewValue = label.LineBreakMode,
-					PlatformViewValue = GetNativeLineBreakMode(handler)
-				};
-			});
-
-			Assert.Equal(xplatLineBreakMode, values.ViewValue);
-			Assert.Equal(expectedValue, values.PlatformViewValue);
-		}
-
-		[Fact(DisplayName = "LineBreakMode does not affect to MaxLines")]
-		public async Task LineBreakModeDoesNotAffectMaxLines()
-		{
-			var label = new LabelStub()
-			{
-				Text = "Lorem ipsum dolor sit amet, consectetur adipiscing elit",
-				MaxLines = 3,
-				LineBreakMode = LineBreakMode.WordWrap,
-			};
-
-			var handler = await CreateHandlerAsync(label);
-			var platformLabel = GetPlatformLabel(handler);
-
-			await InvokeOnMainThreadAsync(() =>
-			{
-				Assert.Equal(3, GetNativeMaxLines(handler));
-				Assert.Equal(LineBreakMode.WordWrap.ToPlatform(), GetNativeLineBreakMode(handler));
-
-				label.LineBreakMode = LineBreakMode.CharacterWrap;
-				platformLabel.UpdateLineBreakMode(label);
-
-				Assert.Equal(3, GetNativeMaxLines(handler));
-				Assert.Equal(LineBreakMode.CharacterWrap.ToPlatform(), GetNativeLineBreakMode(handler));
-			});
-		}
-
-		[Fact(DisplayName = "Single LineBreakMode changes MaxLines")]
-		public async Task SingleLineBreakModeChangesMaxLines()
-		{
-			var label = new LabelStub()
-			{
-				Text = "Lorem ipsum dolor sit amet, consectetur adipiscing elit",
-				MaxLines = 3,
-				LineBreakMode = LineBreakMode.WordWrap,
-			};
-
-			var handler = await CreateHandlerAsync(label);
-			var platformLabel = GetPlatformLabel(handler);
-
-			await InvokeOnMainThreadAsync(() =>
-			{
-				Assert.Equal(3, GetNativeMaxLines(handler));
-				Assert.Equal(LineBreakMode.WordWrap.ToPlatform(), GetNativeLineBreakMode(handler));
-
-				label.LineBreakMode = LineBreakMode.HeadTruncation;
-				platformLabel.UpdateLineBreakMode(label);
-
-				Assert.Equal(1, GetNativeMaxLines(handler));
-				Assert.Equal(LineBreakMode.HeadTruncation.ToPlatform(), GetNativeLineBreakMode(handler));
-			});
-		}
-
-		[Theory(DisplayName = "Unsetting single LineBreakMode resets MaxLines")]
-		[InlineData(LineBreakMode.HeadTruncation)]
-		[InlineData(LineBreakMode.NoWrap)]
-		public async Task UnsettingSingleLineBreakModeResetsMaxLines(LineBreakMode newMode)
-		{
-			var label = new LabelStub()
-			{
-				Text = "Lorem ipsum dolor sit amet, consectetur adipiscing elit",
-				MaxLines = 3,
-				LineBreakMode = LineBreakMode.WordWrap,
-			};
-
-			var handler = await CreateHandlerAsync(label);
-			var platformLabel = GetPlatformLabel(handler);
-
-			await InvokeOnMainThreadAsync(() =>
-			{
-				Assert.Equal(3, GetNativeMaxLines(handler));
-				Assert.Equal(LineBreakMode.WordWrap.ToPlatform(), GetNativeLineBreakMode(handler));
-
-				label.LineBreakMode = newMode;
-				platformLabel.UpdateLineBreakMode(label);
-
-				Assert.Equal(1, GetNativeMaxLines(handler));
-				Assert.Equal(newMode.ToPlatform(), GetNativeLineBreakMode(handler));
-
-				label.LineBreakMode = LineBreakMode.WordWrap;
-				platformLabel.UpdateLineBreakMode(label);
-
-				Assert.Equal(3, GetNativeMaxLines(handler));
-				Assert.Equal(LineBreakMode.WordWrap.ToPlatform(), GetNativeLineBreakMode(handler));
-			});
-		}
-
-		[Fact(DisplayName = "MaxLines Initializes Correctly")]
-		public async Task MaxLinesInitializesCorrectly()
-		{
-			var label = new LabelStub()
-			{
-				Text = "Lorem ipsum dolor sit amet, consectetur adipiscing elit",
-				MaxLines = 2
-			};
-
-			await ValidatePropertyInitValue(label, () => label.MaxLines, GetNativeMaxLines, label.MaxLines);
-		}
-
 		[Fact]
 		[Category(TestCategory.TextFormatting)]
 		public async Task LineHeightAppliedWhenTextAdded()
@@ -430,26 +307,6 @@ namespace Microsoft.Maui.DeviceTests
 
 			Assert.Equal(expectedLineHeight, actualLineHeight);
 			Assert.Equal(expectedCharacterSpacing, actualCharacterSpacing);
-		}
-
-		[Theory(DisplayName = "Negative MaxLines value with wrap is correct")]
-#if __IOS__
-		[InlineData(0)]
-#elif __ANDROID__
-		[InlineData(int.MaxValue)]
-#endif
-		public async Task NegativeMaxValueWithWrapIsCorrect(int expectedLines)
-		{
-			var label = new LabelStub()
-			{
-				Text = "Lorem ipsum dolor sit amet, consectetur adipiscing elit",
-				MaxLines = -1,
-				LineBreakMode = LineBreakMode.WordWrap
-			};
-
-			var platformValue = await GetValueAsync(label, GetNativeMaxLines);
-
-			Assert.Equal(expectedLines, platformValue);
 		}
 
 		[Fact(DisplayName = "LineHeight Initializes Correctly")]
