+++ src/Core/tests/DeviceTests/Handlers/Label/LabelHandlerTests.iOS.cs
@@ -186,9 +186,6 @@ namespace Microsoft.Maui.DeviceTests
 		Color GetNativeTextColor(LabelHandler labelHandler) =>
 			GetPlatformLabel(labelHandler).TextColor.ToColor();
 
-		int GetNativeMaxLines(LabelHandler labelHandler) =>
- 			(int)GetPlatformLabel(labelHandler).Lines;
-
 		double GetNativeCharacterSpacing(LabelHandler labelHandler)
 		{
 			var platformLabel = GetPlatformLabel(labelHandler);
@@ -207,9 +204,6 @@ namespace Microsoft.Maui.DeviceTests
 
 		UITextAlignment GetNativeHorizontalTextAlignment(LabelHandler labelHandler) =>
 			GetPlatformLabel(labelHandler).TextAlignment;
-
-		UILineBreakMode GetNativeLineBreakMode(LabelHandler labelHandler) =>
-			GetPlatformLabel(labelHandler).LineBreakMode;
 
 		double GetNativeLineHeight(LabelHandler labelHandler)
 		{
